feat: replace selects from five to six steps
