+++ src/domain/onboarding/molecules/mobile-onboarding-step-six-startup/mobile-onboarding-step-six-startup.types.ts
@@ -1,4 +1,4 @@
-import { IIndustrialSector, ITeamSize } from '@interfaces';
+import { IIndustrialSector, IStartupSector } from '@interfaces';
 
 import { IMobileOnboardingStepSixStartupData } from '../../onboarding.types';
 
@@ -6,5 +6,5 @@ export interface IMobileOnboardingStepSixStartupProps {
   defaultValues?: IMobileOnboardingStepSixStartupData;
   industrialSectors: IIndustrialSector[];
   onContinueButtonClick: (data: IMobileOnboardingStepSixStartupData) => void;
-  teamSizes: ITeamSize[];
+  startupSectors: IStartupSector[];
 }
