+++ src/infrastructure/i18n/languages/en/onboarding/mobile-onboarding-step-six-startup.translations.ts
@@ -3,10 +3,10 @@ export default {
   'component.mobile.onboarding.step.six.startup.subheading': 'Step 6/5',
   'component.mobile.onboarding.step.six.startup.heading': 'Setup Startup',
 
-  // Team Size
-  'component.mobile.onboarding.step.six.startup.team.size.field.label': 'Team Size',
-
   // Industrial Sectors
   'component.mobile.onboarding.step.six.startup.industrial.sectors.field.label':
     'Industrial Sectors',
+
+  // Sectors
+  'component.mobile.onboarding.step.six.startup.sectors.field.label': 'Sectors',
 };
